mmc: disable a Hynix clock "hack" on u8815
This seems to be bricking devices with a certain Hynix mmc on u8815. :(

These devices die naturally anyway with identical symptoms, and all
replacements seem to be Samsung mmcs... but 3.4 definitely seems to be
accelerating the deaths. As getting a guinea pig is hard/impossible,
we just have to assume this is the cause for now. It's the only stand-out
difference.

Apologies to anyone who lost a device because of this, but there was really
no way to know in advance. The Hynix model seemed rare enough _before_...